+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Added Typebot integration
 * Added proxy endpoint
 * Added send and date_time in webhook data
+* Added support for messaging with ads on chatwoot
 
 ### Fixed